versao semi-final. Ainda precisa de testes
