+++ Paciente_Ler_e_Escrever_ListaEncadeada.h
@@ -1,5 +1,5 @@
 #ifndef PACIENTE_LER_E_ESCREVER_LISTAENCADEADA_H
-#define  PACIENTE_LER_E_ESCREVER_LISTAENCADEADA_H
+#define PACIENTE_LER_E_ESCREVER_LISTAENCADEADA_H
 
 #define NOME_ARQUIVO "dados_pacientes.bin"
